Allow fetch mode to output the plan.json
A common way to use build-env is to fetch all packages,
and then build them offline in a hermetic build environment.

The build step requires the plan.json file, so its convenient for
the fetch step to be able to output it, as opposed to needing to
run the plan step separately.