Proper async support
This patch makes sure we compile each setup script a single time,
even when we are building multiple units in each package. This is
done by introducing another Map of Async operations, keyed on the
package, which we wait on before running the Setup executable of
that package.

We also limit the number of simultaneous builds using a QSem semaphore.